requirements.txt
macfsevents
